feat: share grant editor template
